\name{objectNames}
\alias{objectNames}
\title{Get topojson object names}
\usage{
  objectNames(f)
}
\arguments{
  \item{f}{topojson format file name}
}
\value{
  vector of names
}
\description{
  Get the names of objects in a topojson file
}
\details{
  Get the names of objects
}
\author{
  Barry S Rowlingson
}

