\name{plotpolys}
\alias{plotpolys}
\title{Plots a list of SP Polygons}
\usage{
  plotpolys(polylist, names = c())
}
\arguments{
  \item{polylist}{list of sp 'Polygon'}

  \item{names}{vector of names for the polygons, optional}
}
\value{
  a SpatialPolygons object and a plot
}
\description{
  Plots a list of SP Polygons
}

