\name{readObject}
\alias{readObject}
\title{read an object from a topojson file}
\usage{
  readObject(f, objectname)
}
\arguments{
  \item{f}{topojson file name}

  \item{objectname}{object name}
}
\value{
  SpatialPolygons object
}
\description{
  Read an object from a topojson file
}
\details{
  Reads topojson files
}
\author{
  Barry S Rowlingson
}

